chore: houskeeping, split up files and make code more radable
